feat: add '/about' links
TODO: route to prof / pers tabs
