+++ src/app/page.tsx
@@ -18,7 +18,14 @@ export default function Home() {
         <h2 className="home__description-title gradient-text">Full-Stack Developer</h2>
         <div className="line-break"/>
         <p className="home__description-text">I am Tristan Miller &#x2014; a full-stack developer that loves creating user-accessible web and desktop applications by following common industry standards and best practices.</p>
-        <Link className="home__projects-link" href="/projects">Check out my projects</Link>
+        <Link className="home__page-link" href="/projects">Check out my <u>Projects</u></Link>
+        <div className="home__description-about">
+          <h3 className="home__description-about-text">Click to learn more about my:</h3>
+          <div className="home__description-about-links">
+            <Link className="home__page-link" href="/about"><u>Professional</u> History</Link>
+            <Link className="home__page-link" href="/about"><u>Personal</u> History</Link>
+          </div>
+        </div>
       </div>
       <Languages />
       <Frameworks />
